implemented/[https://trello.com/c/IDjGuHKn/4-actions-handling]/[https://trello.com/c/Eznoh8Bt/3-content-sizing-with-keyboards]

+++ PhotoApp/MainViewController/Timeline/TimeLineViewController.swift
@@ -11,10 +11,10 @@ import UIKit
   struct CategoryModel {
     var id: String 
     var name: String
-    var fred: Float
-    var fgreen: Float
-    var fblue: Float
-    var falpha: Float
+    var fred: CGFloat
+    var fgreen: CGFloat
+    var fblue: CGFloat
+    var falpha: CGFloat
     var isSelected: Int
     
 }
